requirements.txt
PyMuPDF
beautifulsoup4
chonkie
chonkie[semantic]
sentence-transformers
torch
rich
pytest
numpy
pytesseract
Pillow
openai-whisper
qdrant-client
whoosh
neo4j
together
python-dotenv
crewai
deepeval
llm-ie
jupyter